H issue #2063 do delegation calculation just when user clicks delegate button
